security/check if ID user provided really belongs to them
When user provides task id or list id, check their account to find out if it really belongs to them.
User has a list of List models, each List model has a list of Task model objects. When applying changes to either List or Task, check if the list belogs to user first, then additionally, if the Task belongs to the List, that was checked before.